Support remote Deno KV databases
With this change, the `DB_FILE_NAME` environment variable can be set
to the http[s] URL of a Deno KV database. In this case, an additional
variable, `DENO_KV_ACCESS_TOKEN`, must be set to the access token for
the database.

This allows for several instances of the application or even derivative
applications to share the same database.

The `DB_FILE_NAME` environment variable can still be set to a local file
path, as before. This also remains the default.