Implement `bump <n>` CLI command
When restoring a backup, you may need to "bump" the counter of namespaces
to avoid conflicts with ASNs registered after the backup was made.

This commit adds a new CLI command to bump the counter of a namespace.

You can specify the namespace to bump or bump all namespaces.

Use the `stats` command to get a sense of how much you need to bump.